refactor: rename lexicon.py lexicon_helper.py for standard style
